#13 댓글목록-view, dao

+++ jspBoard/src/main/java/org/iclass/dao/CommunityCommentsDao.java
@@ -26,7 +26,7 @@ public class CommunityCommentsDao {
 		return idx;
 	}
 
-	public int delete(int idx) {
+	public int delete(long idx) {
 		SqlSession mapper = SqlSessionBean.getSession();
 		int result = mapper.delete("communityComments.delete", idx);
 		mapper.commit();
